replace todo app with context api
