+++ db.json
@@ -2,13 +2,13 @@
   "todo-list": [
     {
       "text": "adfasdf",
-      "isDone": false,
+      "isDone": true,
       "timeStamp": "12-20-21 12:04",
       "id": 1
     },
     {
       "text": "asdfadsf",
-      "isDone": false,
+      "isDone": true,
       "timeStamp": "12-20-21 12:04",
       "id": 2
     },
@@ -17,6 +17,30 @@
       "isDone": true,
       "timeStamp": "12-20-21 12:04",
       "id": 3
+    },
+    {
+      "text": "asdfdas",
+      "isDone": false,
+      "timeStamp": "12-20-21 01:53",
+      "id": 4
+    },
+    {
+      "text": "asdfdasf",
+      "isDone": false,
+      "timeStamp": "12-20-21 01:54",
+      "id": 5
+    },
+    {
+      "text": "asfdsaf",
+      "isDone": false,
+      "timeStamp": "12-20-21 01:54",
+      "id": 6
+    },
+    {
+      "text": "asdfdasf",
+      "isDone": false,
+      "timeStamp": "12-20-21 01:55",
+      "id": 7
     }
   ]
 }
